Revert "Move Setting related classes into config.flexible.setting package" and "Rename SettingConstraint to Constraint"
This reverts commit 53f2cf76f and 350d5c6f8.
