+++ engine-tests/src/test/java/org/terasology/config/flexible/constraints/NumberRangeConstraintTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.terasology.config.flexible.setting.constraints;
+package org.terasology.config.flexible.constraints;
 
 import org.junit.Test;
 import org.junit.experimental.runners.Enclosed;
