+++ engine/src/main/java/org/terasology/config/flexible/FlexibleConfig.java
@@ -15,7 +15,6 @@
  */
 package org.terasology.config.flexible;
 
-import org.terasology.config.flexible.setting.Setting;
 import org.terasology.engine.SimpleUri;
 
 import java.io.Reader;
